drivers: usb: move DWC2 hardware register header to synopsys directory
This can also be helpful if we need to share code for the relevant
drivers.
