+++ drivers/usb/udc/udc_dwc2_stm32f4_fsotg.h
@@ -8,7 +8,7 @@
 
 #include <zephyr/sys/sys_io.h>
 #include <zephyr/drivers/clock_control/stm32_clock_control.h>
-#include <usb_dwc2_hw.h>
+#include <synopsys/usb_dwc2_hw.h>
 
 struct usb_dw_stm32_clk {
 	const struct device *const dev;
